GetPQueue function which make a priority queue, insert messages with priorities to the messages_processing queue

+++ dispatcher/dispatcher.go
@@ -10,7 +10,7 @@ type Notification struct {
 	UserID     string
 	CreatedAt  int64
 	SendBefore time.Duration
-	Priority   int
+	Priority   uint8
 	Counter    int
 }
